+++ 3-Express-Server/.gitignore
@@ -18,6 +18,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.prettierrc
 
 npm-debug.log*
 yarn-debug.log*
