+++ 3-Express-Server/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "challenge-express-server",
+  "version": "1.0.0",
+  "description": "",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/LucasKoval/CH-Server-Express.git"
+  },
+  "author": "Lucas Koval",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/LucasKoval/CH-Server-Express/issues"
+  },
+  "homepage": "https://github.com/LucasKoval/CH-Server-Express#readme",
+  "dependencies": {
+    "express": "^4.17.3"
+  }
+}
