+++ 3-Express-Server/src/index.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Express Server</title>
+  </head>
+  <body style="background-color: grey">
+    <div style="display: flex; justify-content: center; align-items: center">
+      <h1
+        style="
+          font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu,
+            Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+          display: flex;
+          color: white;
+        "
+      >
+        CH Challenge - Express Server
+      </h1>
+    </div>
+    <div style="display: flex; justify-content: center; align-items: center">
+      <a href="/productos" style="margin: 0 2rem">
+        <button
+          type="button"
+          style="font-size: 1rem; padding: 1rem; background-color: black; color: white"
+        >
+          All Products
+        </button></a
+      >
+      <a href="/productoRandom" style="margin: 0 2rem">
+        <button
+          type="button"
+          style="font-size: 1rem; padding: 1rem; background-color: black; color: white"
+        >
+          Random Product
+        </button></a
+      >
+    </div>
+  </body>
+</html>
