--- 3-Template-Engines/EJS/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "challenge-template-engines-ejs",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/app.js",
-    "dev": "nodemon src/app.js"
-  },
-  "author": "Lucas Koval",
-  "license": "ISC",
-  "dependencies": {
-    "ejs": "^3.1.6",
-    "express": "^4.17.3"
-  }
-}
