--- 3-Template-Engines/EJS/public/css/styles.css
@@ -1,123 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-body {
-  width: 100%;
-  height: auto;
-  min-height: 100vh;
-  background: linear-gradient(-225deg, #2c5364, #203a43, #0f2027);
-  padding: 0;
-  margin: 0;
-}
-h1,
-p,
-span,
-input,
-button,
-div,
-a {
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell,
-    Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-}
-
-h1 {
-  font-size: 3rem;
-  padding-top: 1rem;
-  text-align: center;
-  text-transform: uppercase;
-  font-family: 'Staatliches', cursive;
-  letter-spacing: 1px;
-  color: white;
-}
-
-h2 {
-  font-size: 2rem;
-  padding-top: 3rem;
-  text-align: center;
-  text-transform: uppercase;
-  font-family: 'Staatliches', cursive;
-  letter-spacing: 1px;
-  color: white;
-}
-
-label {
-  font-size: 18px;
-  color: white;
-  font-weight: bold;
-  font-style: italic;
-}
-
-input {
-  width: 100%;
-  height: 38px;
-  padding: 6px 10px;
-  margin-top: 0.2rem;
-  margin-bottom: 2rem;
-  background-color: #fff;
-  border: 1px solid #d1d1d1;
-  border-radius: 4px;
-  box-shadow: none;
-  box-sizing: border-box;
-}
-input:focus {
-  outline: none;
-  border: 1px solid white;
-}
-
-button {
-  width: 100%;
-  height: 40px;
-  border-radius: 5px;
-  margin-top: 20px;
-  padding: 0 2rem;
-  color: #ffffff;
-  background: #00b4db;
-  border: none;
-  align-self: flex-end;
-  font-style: normal;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 20px;
-  text-align: center;
-  box-sizing: border-box;
-  cursor: pointer;
-}
-button:hover {
-  background: #009dc0;
-  transition: all 0.4s ease;
-}
-
-.main {
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  padding-top: 2rem;
-  padding-bottom: 2rem;
-}
-
-.column {
-  width: 80%;
-  max-width: 500px;
-  display: flex;
-  flex-flow: column;
-  justify-content: flex-start;
-  align-items: center;
-  align-self: baseline;
-}
-
-.message {
-  background-color: rgb(250, 216, 216);
-  padding: 1rem 2rem;
-  border-radius: 8px;
-  border: 1px solid rgb(216, 0, 0);
-}
-.message h2 {
-  color: rgb(216, 0, 0);
-  padding: 0;
-  margin: 0;
-}
-
-.image-container img {
-  width: 56px;
-}
