--- 3-Template-Engines/EJS/src/views/index.ejs
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" href="/icon/nodejs.ico" />
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-      crossorigin="anonymous"
-    />
-    <link href="https://fonts.googleapis.com/css?family=Staatliches" rel="stylesheet" />
-    <link href="/css/styles.css" rel="stylesheet" />
-    <title>Template Engines - EJS</title>
-  </head>
-  <body>
-    <div class="main">
-      <div class="column">
-        <h1>CH - Template Engines (EJS)</h1>
-
-        <a href="/productos">
-          <button type="button" style="background-color: #ff6962">See All Products</button>
-        </a>
-
-        <h2>Enter new Product</h2>
-
-        <form action="/productos" method="POST">
-          <label>Name</label>
-          <input type="text" name="title" placeholder="Example: Notebook Asus" required />
-
-          <label>Price</label>
-          <input type="number" name="price" placeholder="Example: 1500" required />
-
-          <label>Image URL</label>
-          <input type="text" name="thumbnail" placeholder="Product image URL" />
-
-          <button type="submit">Add Product</button>
-        </form>
-      </div>
-    </div>
-  </body>
-</html>
