--- 3-Template-Engines/EJS/src/views/products.ejs
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" href="/icon/nodejs.ico" />
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-      crossorigin="anonymous"
-    />
-    <link href="https://fonts.googleapis.com/css?family=Staatliches" rel="stylesheet" />
-    <link href="/css/styles.css" rel="stylesheet" />
-    <title>Template Engines - EJS</title>
-  </head>
-  <body>
-    <div class="main">
-      <div class="column">
-        <h1>CH - Template Engines (EJS)</h1>
-
-        <a href="/">
-          <button type="button" style="background-color: #ff6962">Back to Home</button>
-        </a>
-
-        <h2 class="mb-4">Product View</h2>
-
-        <% if(allProducts.length > 0){ %>
-        <table class="table table-dark table-striped">
-          <thead>
-            <tr>
-              <th scope="col" class="ps-4">Name</th>
-              <th scope="col" class="text-center">Price</th>
-              <th scope="col" class="text-center">Image</th>
-            </tr>
-          </thead>
-          <tbody>
-            <% for (const product of allProducts) {%>
-            <tr>
-              <th scope="row" class="ps-4 align-middle"><%=product.title%></th>
-              <td class="text-center align-middle"><%=product.price%></td>
-              <td>
-                <div class="image-container text-center">
-                  <img src="<%= product.thumbnail %>" alt="Product image" class="img-thumbnail" />
-                </div>
-              </td>
-            </tr>
-            <%}%>
-          </tbody>
-        </table>
-        <% } else { %>
-        <div class="message">
-          <h2>There are no products</h2>
-        </div>
-        <% } %>
-      </div>
-    </div>
-  </body>
-</html>
