--- 3-Template-Engines/Handlebars/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "challenge-template-engines-handlebars",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/app.js",
-    "dev": "nodemon src/app.js"
-  },
-  "author": "Lucas Koval",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.17.3",
-    "express-handlebars": "^6.0.5"
-  }
-}
