--- 3-Template-Engines/Handlebars/src/database/products.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": 1,
-    "title": "Escuadra",
-    "price": 123.45,
-    "thumbnail": "https://cdn3.iconfinder.com/data/icons/education-209/64/ruler-triangle-stationary-school-64.png"
-  },
-  {
-    "id": 2,
-    "title": "Calculadora",
-    "price": 234.56,
-    "thumbnail": "https://cdn3.iconfinder.com/data/icons/education-209/64/calculator-math-tool-school-64.png"
-  },
-  {
-    "id": 3,
-    "title": "Globo Terráqueo",
-    "price": 345.67,
-    "thumbnail": "https://cdn3.iconfinder.com/data/icons/education-209/64/globe-earth-geograhy-planet-school-64.png"
-  }
-]
