--- 3-Template-Engines/Handlebars/src/views/form.hbs
@@ -1,18 +1,0 @@
-<a href='/productos'>
-  <button type='button' style='background-color: #ff6962'>See All Products</button>
-</a>
-
-<h2>Enter new Product</h2>
-
-<form action='/productos' method='POST'>
-  <label>Name</label>
-  <input type='text' name='title' placeholder='Example: Notebook Asus' required />
-
-  <label>Price</label>
-  <input type='number' name='price' placeholder='Example: 1500' required />
-
-  <label>Image URL</label>
-  <input type='text' name='thumbnail' placeholder='Product image URL' />
-
-  <button type='submit'>Add Product</button>
-</form>
