--- 3-Template-Engines/Handlebars/src/views/products.hbs
@@ -1,34 +1,0 @@
-<a href='/'>
-  <button type='button' style='background-color: #ff6962'>Back to Home</button>
-</a>
-
-<h2 class='mb-4'>Product View</h2>
-
-{{#if allProducts}}
-  <table class='table table-dark table-striped'>
-    <thead>
-      <tr>
-        <th scope='col' class='ps-4'>Name</th>
-        <th scope='col' class='text-center'>Price</th>
-        <th scope='col' class='text-center'>Image</th>
-      </tr>
-    </thead>
-    <tbody>
-      {{#each allProducts}}
-        <tr>
-          <th scope='row' class='ps-4 align-middle'>{{this.title}}</th>
-          <td class='text-center align-middle'>{{this.price}}</td>
-          <td>
-            <div class='image-container text-center'>
-              <img src={{this.thumbnail}} alt='Product image' class='img-thumbnail' />
-            </div>
-          </td>
-        </tr>
-      {{/each}}
-    </tbody>
-  </table>
-{{else}}
-  <div class='message'>
-    <h2>There are no products</h2>
-  </div>
-{{/if}}
