--- 3-Template-Engines/Pug/src/views/form.pug
@@ -1,14 +1,0 @@
-extends layout
-
-block content
-  a(href='/productos')
-    button(type='button' style='background-color: #ff6962') See All Products
-  h2 Enter new Product
-  form(action='/productos' method='POST')
-    label Name
-    input(type='text' name='title' placeholder='Example: Notebook Asus' required='')
-    label Price
-    input(type='number' name='price' placeholder='Example: 1500' required='')
-    label Image URL
-    input(type='text' name='thumbnail' placeholder='Product image URL')
-    button(type='submit') Add Product
