--- 3-Template-Engines/Pug/src/views/layout.pug
@@ -1,20 +1,0 @@
-html(lang='en')
-  head
-    meta(charset='UTF-8')
-    meta(http-equiv='X-UA-Compatible' content='IE=edge')
-    meta(name='viewport' content='width=device-width, initial-scale=1.0')
-    link(rel='icon' href='/icon/nodejs.ico')
-    link(href='https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css' rel='stylesheet' integrity='sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3' crossorigin='anonymous')
-    link(href='https://fonts.googleapis.com/css?family=Staatliches' rel='stylesheet')
-    link(href='/css/styles.css' rel='stylesheet')
-    title Template Engines - Pug
-  body
-    .main
-      .column
-        h1 CH - Template Engines (Pug)
-        block content
-
-
-
-
-
