--- 5-Template-Engines/README.md
@@ -1,5 +1,0 @@
-# Challenge CH - Template Engines
-
-Para el proyecto final elijo trabajar con EJS porque me parece el más versátil y fácil de trabajar.
-El código JS que se inyecta en el HTML es el mismo que cualquier código JS y no tiene funciones específicas del motor, como si lo tiene Handlebars.
-Y Pug es un asco.
